// libs/prisma/generated/prisma/models/Currency.ts
export type CurrencyCreateInput = {
  id?: string
  name: string
  symbol: $Enums.CurrencyAsset
  description?: string | null
  createdAt?: Date | string
  updatedAt?: Date | string
  balance?: Prisma.BalanceCreateNestedManyWithoutCurrencyInput
}